+++ src/main/java/at/fhburgenland/service/FahrzeugtypService.java
@@ -1,23 +1,24 @@
 package at.fhburgenland.service;
 
+import at.fhburgenland.model.Fahrer;
+import at.fhburgenland.model.Fahrzeug;
 import at.fhburgenland.model.Fahrzeugtyp;
-import at.fhburgenland.model.Rennstrecke;
+import at.fhburgenland.model.Nationalitaet;
 import jakarta.persistence.*;
 
 import java.util.List;
-import java.util.Scanner;
+
 
 public class FahrzeugtypService {
     private static EntityManagerFactory EMF = Persistence.createEntityManagerFactory("project");
-    private static Scanner scanner = new Scanner(System.in);
 
-    public static void fahrzeugtypHinzufuegen(String modell, String motor, int gewichtKg) {
+
+    public static void fahrzeugtypHinzufuegen(Fahrzeugtyp fahrzeugtyp) {
         EntityManager em = EMF.createEntityManager();
         EntityTransaction et = null;
         try {
             et = em.getTransaction();
             et.begin();
-            Fahrzeugtyp fahrzeugtyp = new Fahrzeugtyp(modell, motor, gewichtKg);
             System.out.println("Neuer Fahrzeugtyp wurde angelegt: " + fahrzeugtyp);
             em.persist(fahrzeugtyp);
             et.commit();
@@ -30,7 +31,7 @@ public class FahrzeugtypService {
         }
     }
 
-    public static void alleFahrzeugtypenAnzeigen() {
+    public static List<Fahrzeugtyp> alleFahrzeugtypenAnzeigen() {
         EntityManager em = EMF.createEntityManager();
         String query = "SELECT ft FROM Fahrzeugtyp ft";
         TypedQuery<Fahrzeugtyp> tq = em.createQuery(query, Fahrzeugtyp.class);
@@ -39,26 +40,27 @@ public class FahrzeugtypService {
 
         try {
             fahrzeugtypListe = tq.getResultList();
-            for (Fahrzeugtyp fahrzeugtyp : fahrzeugtypListe) {
-                System.out.println("Fahrzeugtyp Nr: " + fahrzeugtyp.getFahrzeugtypId() + ", Modell " + fahrzeugtyp.getModell() + ", Motor " + fahrzeugtyp.getMotor() + ", Gewicht " + fahrzeugtyp.getGewichtKg() + "kg");
-            }
         } catch (Exception e) {
             e.printStackTrace();
         } finally {
             em.close();
         }
+        return fahrzeugtypListe;
     }
 
-    public static Fahrzeugtyp fahrzeugtypAuswaehlen() {
-        alleFahrzeugtypenAnzeigen();
-        System.out.println("Bitte gib die ID des gewünschten Fahrzeugtyps ein:");
-        int fahrzeugtypId = scanner.nextInt();
-        scanner.nextLine();
+    public static Fahrzeugtyp fahrzeugtypAnzeigenNachId(int fahrzeugtypId) {
         EntityManager em = EMF.createEntityManager();
-        Fahrzeugtyp ausgewaehlterFahrzeugtyp = null;
+        Fahrzeugtyp fahrzeugtyp = null;
+
         try {
-            ausgewaehlterFahrzeugtyp = em.find(Fahrzeugtyp.class, fahrzeugtypId);
-            if (ausgewaehlterFahrzeugtyp == null) {
+            fahrzeugtyp = em.find(Fahrzeugtyp.class, fahrzeugtypId);
+            if (fahrzeugtyp != null) {
+                System.out.println("Fahrzeugtyp gefunden:");
+                System.out.println("ID: " + fahrzeugtyp.getFahrzeugtypId());
+                System.out.println("Modell: " + fahrzeugtyp.getModell());
+                System.out.println("Motor: " + fahrzeugtyp.getMotor());
+                System.out.println("Gewicht: " + fahrzeugtyp.getGewichtKg());
+            } else {
                 System.err.println("Kein Fahrzeugtyp mit dieser ID gefunden.");
             }
         } catch (Exception e) {
@@ -66,25 +68,50 @@ public class FahrzeugtypService {
         } finally {
             em.close();
         }
-        return ausgewaehlterFahrzeugtyp;
+        return fahrzeugtyp;
+    }
+
+    public static void fahrzeugtypUpdaten(Fahrzeugtyp fahrzeugtyp) {
+        EntityManager em = EMF.createEntityManager();
+        EntityTransaction et = null;
+
+        try {
+            et = em.getTransaction();
+            et.begin();
+
+            em.merge(fahrzeugtyp);
+            et.commit();
+            System.out.println("Fahrzeugtyp erfolgreich aktualisiert: " + fahrzeugtyp);
+
+        } catch (Exception e) {
+            if (et != null) {
+                et.rollback();
+            }
+            e.printStackTrace();
+        } finally {
+            em.close();
+        }
     }
 
     public static void fahrzeugtypLoeschen(int fahrzeugtypId) {
         EntityManager em = EMF.createEntityManager();
         EntityTransaction et = null;
-        Fahrzeugtyp fahrzeugtyp = null;
 
         try {
             et = em.getTransaction();
             et.begin();
-            fahrzeugtyp = em.find(Fahrzeugtyp.class, fahrzeugtypId);
-            em.remove(fahrzeugtyp);
-            et.commit();
-            System.out.format("Fahrzeugtyp %d erfolgreich gelöscht.\n", fahrzeugtypId);
-        } catch (Exception e) {
+            Fahrzeugtyp fahrzeugtyp= em.find(Fahrzeugtyp.class, fahrzeugtypId);
+            if (fahrzeugtyp != null) {
+                em.remove(fahrzeugtyp);
+                et.commit();
+            } else {
+                System.err.println("Fahrzeugtyp nicht gefunden.");
+            }
+        }catch (Exception e) {
             if (et != null) {
                 et.rollback();
-                System.out.println(e.getMessage());
+                System.out.println("Fehler beim Löschen des Fahrzeugtyps.");
+                e.printStackTrace();
             }
         } finally {
             em.close();
