+++ src/main/java/at/fhburgenland/service/RennenService.java
@@ -1,22 +1,21 @@
 package at.fhburgenland.service;
 
+import at.fhburgenland.model.Fahrzeug;
+import at.fhburgenland.model.Nationalitaet;
 import at.fhburgenland.model.Rennen;
-import at.fhburgenland.model.Rennstrecke;
 import jakarta.persistence.*;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 public class RennenService {
     private static EntityManagerFactory EMF = Persistence.createEntityManagerFactory("project");
 
-    public static void rennenHinzufuegen(LocalDateTime datumUhrzeit, Rennstrecke rennstrecke) {
+    public static void rennenHinzufuegen(Rennen rennen) {
         EntityManager em = EMF.createEntityManager();
         EntityTransaction et = null;
         try {
             et = em.getTransaction();
             et.begin();
-            Rennen rennen = new Rennen(datumUhrzeit, rennstrecke);
             System.out.println("Neues Rennen wurde angelegt: " + rennen);
             em.persist(rennen);
             et.commit();
@@ -29,7 +28,7 @@ public class RennenService {
         }
     }
 
-    public static void alleRennenAnzeigen() {
+    public static List<Rennen> alleRennenAnzeigen() {
         EntityManager em = EMF.createEntityManager();
         String query = "SELECT r FROM Rennen r";
         TypedQuery<Rennen> tq = em.createQuery(query, Rennen.class);
@@ -38,10 +37,52 @@ public class RennenService {
 
         try {
             rennenListe = tq.getResultList();
-            for (Rennen rennen : rennenListe) {
-                System.out.println("Rennen Nr: " + rennen.getRennenId() + ", Datum und Uhrzeit " + rennen.getDatumUhrzeit() + ", " + rennen.getRennstrecke());
+        } catch (Exception e) {
+            e.printStackTrace();
+        } finally {
+            em.close();
+        }
+        return rennenListe;
+    }
+
+    public static Rennen rennenAnzeigenNachId(int rennenId) {
+        EntityManager em = EMF.createEntityManager();
+        Rennen rennen = null;
+
+        try {
+            rennen = em.find(Rennen.class, rennenId);
+            if (rennen != null) {
+                System.out.println("Rennen gefunden:");
+                System.out.println("ID: " + rennen.getRennenId());
+                System.out.println("Datum und Uhrzeit: " + rennen.getDatumUhrzeit());
+                System.out.println("Rennstrecke: " + rennen.getRennstrecke());
+            } else {
+                System.err.println("Kein Rennen mit dieser ID gefunden.");
             }
         } catch (Exception e) {
+            e.printStackTrace();
+        } finally {
+            em.close();
+        }
+        return rennen;
+    }
+
+    public static void rennenUpdaten(Rennen rennen) {
+        EntityManager em = EMF.createEntityManager();
+        EntityTransaction et = null;
+
+        try {
+            et = em.getTransaction();
+            et.begin();
+
+            em.merge(rennen);
+            et.commit();
+            System.out.println("Rennen erfolgreich aktualisiert: " + rennen);
+
+        } catch (Exception e) {
+            if (et != null) {
+                et.rollback();
+            }
             e.printStackTrace();
         } finally {
             em.close();
@@ -51,19 +92,22 @@ public class RennenService {
     public static void rennenLoeschen(int rennenId) {
         EntityManager em = EMF.createEntityManager();
         EntityTransaction et = null;
-        Rennen rennen = null;
 
         try {
             et = em.getTransaction();
             et.begin();
-            rennen = em.find(Rennen.class, rennenId);
-            em.remove(rennen);
-            et.commit();
-            System.out.format("Rennen %d erfolgreich gelöscht.\n", rennenId);
+            Rennen rennen = em.find(Rennen.class, rennenId);
+            if (rennen != null) {
+                em.remove(rennen);
+                et.commit();
+            } else {
+                System.err.println("Rennen nicht gefunden.");
+            }
         } catch (Exception e) {
             if (et != null) {
                 et.rollback();
-                System.out.println(e.getMessage());
+                System.out.println("Fehler beim Löschen des Rennens.");
+                e.printStackTrace();
             }
         } finally {
             em.close();
