+++ src/main/java/at/fhburgenland/service/TeamService.java
@@ -1,25 +1,20 @@
 package at.fhburgenland.service;
 
-import at.fhburgenland.model.Fahrzeugtyp;
-import at.fhburgenland.model.Hauptsponsor;
-import at.fhburgenland.model.Nationalitaet;
-import at.fhburgenland.model.Team;
+import at.fhburgenland.model.*;
 import jakarta.persistence.*;
 
 import java.util.List;
-import java.util.Scanner;
+
 
 public class TeamService {
     private static EntityManagerFactory EMF = Persistence.createEntityManagerFactory("project");
-    private static Scanner scanner = new Scanner(System.in);
 
-    public static void teamHinzufuegen(String teamName, int gruendungsjahr, Nationalitaet nationalitaet, Hauptsponsor hauptsponsor) {
+    public static void teamHinzufuegen(Team team) {
         EntityManager em = EMF.createEntityManager();
         EntityTransaction et = null;
         try {
             et = em.getTransaction();
             et.begin();
-            Team team = new Team(teamName, gruendungsjahr, nationalitaet, hauptsponsor);
             System.out.println("Neues Team wurde angelegt: " + team);
             em.persist(team);
             et.commit();
@@ -32,7 +27,7 @@ public class TeamService {
         }
     }
 
-    public static void alleTeamsAnzeigen() {
+    public static List<Team> alleTeamsAnzeigen() {
         EntityManager em = EMF.createEntityManager();
         String query = "SELECT t FROM Team t";
         TypedQuery<Team> tq = em.createQuery(query, Team.class);
@@ -41,26 +36,28 @@ public class TeamService {
 
         try {
             teamsListe = tq.getResultList();
-            for (Team team : teamsListe) {
-                System.out.println("Team Nr: " + team.getTeamId() + ", " + team.getTeamName() + ", Gründungsjahr " + team.getGruendungsjahr() + ", " + team.getNationalitaet() + ", " + team.getHauptsponsor());
-            }
         } catch (Exception e) {
             e.printStackTrace();
         } finally {
             em.close();
         }
+        return teamsListe;
     }
 
-    public static Team teamAuswaehlen() {
-        alleTeamsAnzeigen();
-        System.out.println("Bitte gib die ID des gewünschten Teams ein:");
-        int teamId = scanner.nextInt();
-        scanner.nextLine();
+    public static Team teamAnzeigenNachId(int teamId) {
         EntityManager em = EMF.createEntityManager();
-        Team ausgewaehltesTeam = null;
+        Team team = null;
+
         try {
-            ausgewaehltesTeam = em.find(Team.class, teamId);
-            if (ausgewaehltesTeam == null) {
+            team = em.find(Team.class, teamId);
+            if (team != null) {
+                System.out.println("Team gefunden:");
+                System.out.println("ID: " + team.getTeamId());
+                System.out.println("Teamname: " + team.getTeamName());
+                System.out.println("Gründungsjahr: " + team.getGruendungsjahr());
+                System.out.println("Hauptsponsor: " + team.getHauptsponsor());
+                System.out.println("Nationalität: " + team.getNationalitaet());
+            } else {
                 System.err.println("Kein Team mit dieser ID gefunden.");
             }
         } catch (Exception e) {
@@ -68,25 +65,50 @@ public class TeamService {
         } finally {
             em.close();
         }
-        return ausgewaehltesTeam;
+        return team;
+    }
+
+    public static void teamUpdaten(Team team) {
+        EntityManager em = EMF.createEntityManager();
+        EntityTransaction et = null;
+
+        try {
+            et = em.getTransaction();
+            et.begin();
+
+            em.merge(team);
+            et.commit();
+            System.out.println("Team erfolgreich aktualisiert: " + team);
+
+        } catch (Exception e) {
+            if (et != null) {
+                et.rollback();
+            }
+            e.printStackTrace();
+        } finally {
+            em.close();
+        }
     }
 
     public static void teamLoeschen(int teamId) {
         EntityManager em = EMF.createEntityManager();
         EntityTransaction et = null;
-        Team team = null;
 
         try {
             et = em.getTransaction();
             et.begin();
-            team = em.find(Team.class, teamId);
-            em.remove(team);
-            et.commit();
-            System.out.format("Team %d erfolgreich gelöscht.\n", teamId);
+           Team team = em.find(Team.class, teamId);
+            if (team != null) {
+                em.remove(team);
+                et.commit();
+            } else {
+                System.err.println("Team nicht gefunden.");
+            }
         } catch (Exception e) {
             if (et != null) {
                 et.rollback();
-                System.out.println(e.getMessage());
+                System.out.println("Fehler beim Löschen des Teams.");
+                e.printStackTrace();
             }
         } finally {
             em.close();
